Fibonacci class ready

+++ Fibonacci.py
@@ -1,3 +1,3 @@
 from math import sqrt
 def Fibonacci(n):
-	return ((1+sqrt(5))**n-(1-sqrt(5))**n)/(2**n*sqrt(5))
+	return int(((1+sqrt(5))**n-(1-sqrt(5))**n)/(2**n*sqrt(5)))
